@@ -436,9 +453,11 @@
     "mocha": "^8.4.0",
     "prettier": "^2.1.2",
     "pretty-quick": "^3.1.0",
+    "sinon": "^12.0.1",
     "standard-version": "^9.3.1",
+    "stub-spawn-once": "^2.3.0",
     "ts-loader": "^9.2.2",
-    "typescript": "^4.3.2",
+    "typescript": "^4.5.2",
     "vsce": "^1.96.1",
     "vscode-extension-tester": "^4.1.2",
     "vscode-test": "^1.5.2",
