@@ -1,6 +1,6 @@
 import * as fs from 'fs';
 import * as path from 'path';
-import stripIndent from 'strip-indent';
+import { stripIndent } from 'common-tags';
 import handlebars from 'handlebars';
 
 import {
@@ -24,7 +24,7 @@ import {
   sortLineSelections,
   rangeFromStringDefinition,
   escapeEndOfLineForCsv,
-  standardizeFilename,
+  relativeToWorkspace,
   splitStringDefinition,
 } from './utils/workspace-util';
 import { ReviewFileExportSection, GroupBy, ExportFormat, ExportMap, Group } from './interfaces';
@@ -32,7 +32,7 @@ import { CsvEntry, CsvStructure } from './model';
 import { CommentListEntry } from './comment-list-entry';
 import { FileGenerator } from './file-generator';
 import { themeColorForPriority } from './utils/editor-utils';
-const gitCommitId = require('git-commit-id');
+import { gitRevision } from './vcs-provider';
 
 export class ExportFactory {
   private defaultFileName = 'code-review';
@@ -51,7 +51,7 @@ export class ExportFactory {
    */
   private isCommentEligible(entry: CsvEntry): boolean {
     return (
-      (this.currentCommitId === null || entry.sha === this.currentCommitId) &&
+      (this.currentCommitId === null || entry.revision === this.currentCommitId) &&
       (this.currentFilename === null || entry.filename === this.currentFilename)
     );
   }
@@ -114,13 +114,13 @@ export class ExportFactory {
           const title = row.title ? row.title.substring(0, 255) : descShort;
           const fileRow = row.url ? `- file: [${row.filename}](${row.url})${EOL}` : `${row.filename}${EOL}`;
           const linesRow = `- lines: ${row.lines}${EOL}`;
-          const shaRow = row.sha ? `- SHA: ${row.sha}${EOL}${EOL}` : '';
+          const revRow = row.revision ? `- SHA: ${row.revision}${EOL}${EOL}` : '';
           const commentSection = `## Comment${EOL}${row.comment}${EOL}`;
           const additional = row.additional ? `## Additional information${EOL}${row.additional}${EOL}` : '';
           const priority = row.priority ? `## Priority${EOL}${this.priorityName(row.priority)}${EOL}${EOL}` : '';
           const category = row.category ? `## Category${EOL}${row.category}${EOL}${EOL}` : '';
           const code = row.code ? `${EOL}## Source Code${EOL}${EOL}\`\`\`${EOL}${row.code}\`\`\`${EOL}` : '';
-          const description = `${priority}${category}## Affected${EOL}${fileRow}${linesRow}${shaRow}${commentSection}${EOL}${additional}${code}`;
+          const description = `${priority}${category}## Affected${EOL}${fileRow}${linesRow}${revRow}${commentSection}${EOL}${additional}${code}`;
           fs.appendFileSync(outputFile, `"[code review] ${title}","${description}"${EOL}`);
           return row;
         },
@@ -148,14 +148,14 @@ export class ExportFactory {
 
           const fileRow = row.url ? `- file: [${row.filename}](${row.url})${EOL}` : `${row.filename}${EOL}`;
           const linesRow = `- lines: ${row.lines}${EOL}`;
-          const shaRow = row.sha ? `- SHA: ${row.sha}${EOL}${EOL}` : '';
+          const revRow = row.revision ? `- SHA: ${row.revision}${EOL}${EOL}` : '';
           const commentSection = `## Comment${EOL}${row.comment}${EOL}`;
           const additional = row.additional ? `## Additional information${EOL}${row.additional}${EOL}` : '';
           const priority = row.priority ? `## Priority${EOL}${this.priorityName(row.priority)}${EOL}${EOL}` : '';
           const category = row.category ? `## Category${EOL}${row.category}${EOL}${EOL}` : '';
           const code = row.code ? `${EOL}## Source Code${EOL}${EOL}\`\`\`${EOL}${row.code}\`\`\`${EOL}` : '';
 
-          const description = `${priority}${category}## Affected${EOL}${fileRow}${linesRow}${shaRow}${commentSection}${EOL}${additional}${code}`;
+          const description = `${priority}${category}## Affected${EOL}${fileRow}${linesRow}${revRow}${commentSection}${EOL}${additional}${code}`;
 
           fs.appendFileSync(outputFile, `"[code review] ${title}","${description}","code-review","open",""${EOL}`);
           return row;
@@ -173,7 +173,7 @@ export class ExportFactory {
         writeFileHeader: (outputFile: string) => {
           fs.writeFileSync(
             outputFile,
-            `Summary,Description,Priority,sha,filename,url,lines,title,category,comment,additional${EOL}`,
+            `Summary,Description,Priority,revision,filename,url,lines,title,category,comment,additional${EOL}`,
           );
         },
         handleData: (outputFile: string, row: CsvEntry): CsvEntry => {
@@ -187,17 +187,17 @@ export class ExportFactory {
 
           const fileRow = row.url ? `* file: [${row.filename}|${row.url}]${EOL}` : `${row.filename}${EOL}`;
           const linesRow = `* lines: ${row.lines}${EOL}`;
-          const shaRow = row.sha ? `* SHA: ${row.sha}${EOL}${EOL}` : '';
+          const revRow = row.revision ? `* SHA: ${row.revision}${EOL}${EOL}` : '';
           const categorySection = `h2. Category${EOL}${row.category}${EOL}${EOL}`;
           const commentSection = `h2. Comment${EOL}${row.comment}${EOL}`;
           const additional = row.additional ? `h2. Additional information${EOL}${row.additional}${EOL}` : '';
           const code = row.code ? `${EOL}h2. Source Code${EOL}${EOL}{code}${EOL}${row.code}{code}${EOL}` : '';
 
-          const description = `h2. Affected${EOL}${fileRow}${linesRow}${shaRow}${categorySection}${commentSection}${EOL}${additional}${code}`;
+          const description = `h2. Affected${EOL}${fileRow}${linesRow}${revRow}${categorySection}${commentSection}${EOL}${additional}${code}`;
 
           fs.appendFileSync(
             outputFile,
-            `"[code review] ${title}","${description}","${this.priorityName(row.priority)}","${row.sha}","${
+            `"[code review] ${title}","${description}","${this.priorityName(row.priority)}","${row.revision}","${
               row.filename
             }","${row.url}","${row.lines}","${row.title}","${row.category}","${row.comment}","${row.additional}"${EOL}`,
           );
@@ -479,17 +479,17 @@ export class ExportFactory {
   public setFilterByCommit(state: boolean): boolean {
     this.filterByCommit = state;
     if (this.filterByCommit) {
-      try {
-        const gitDirectory = workspace.getConfiguration().get('code-review.gitDirectory') as string;
-        const gitRepositoryPath = path.resolve(this.workspaceRoot, gitDirectory);
+      gitRevision('.', this.workspaceRoot).then(
+        (revision: string) => {
+          this.currentCommitId = revision;
+        },
+        (error: string) => {
+          this.filterByCommit = false;
+          this.currentCommitId = null;
 
-        this.currentCommitId = gitCommitId({ cwd: gitRepositoryPath });
-      } catch (error) {
-        this.filterByCommit = false;
-        this.currentCommitId = null;
-
-        console.log('Not in a git repository. Disabling filter by commit', error);
-      }
+          console.log('Not in a git repository. Disabling filter by commit.', error);
+        },
+      );
     } else {
       this.currentCommitId = null;
     }
@@ -521,7 +521,7 @@ export class ExportFactory {
     if (this.filterByFilename) {
       let filename = window.activeTextEditor?.document.fileName;
       if (filename) {
-        filename = standardizeFilename(this.workspaceRoot, filename);
+        filename = relativeToWorkspace(this.workspaceRoot, filename);
         if (this.currentFilename !== filename) {
           changedFile = true;
           this.currentFilename = filename;
