@@ -7,12 +7,11 @@ import {
   removeTrailingSlash,
   startLineNumberFromStringDefinition,
   endLineNumberFromStringDefinition,
-  standardizeFilename,
+  relativeToWorkspace,
 } from './utils/workspace-util';
-import { CommentListEntry } from './comment-list-entry';
 import { getSelectionStringDefinition, hasSelection } from './utils/editor-utils';
 import { getCsvFileLinesAsArray, setCsvFileLines } from './utils/storage-utils';
-import path from 'path';
+import { revision } from './vcs-provider';
 
 export class ReviewCommentService {
   constructor(private reviewFile: string, private workspaceRoot: string) {}
@@ -29,9 +28,11 @@ export class ReviewCommentService {
       return;
     }
 
-    comment.filename = standardizeFilename(this.workspaceRoot, editor!.document.fileName);
+    comment.filename = relativeToWorkspace(this.workspaceRoot, editor!.document.fileName);
 
-    setCsvFileLines(this.reviewFile, [CsvStructure.formatAsCsvLine(this.finalizeComment(comment))], false);
+    this.finalizeComment(comment, this.workspaceRoot).then((entry: CsvEntry) => {
+      setCsvFileLines(this.reviewFile, [CsvStructure.formatAsCsvLine(entry)], false);
+    });
   }
 
   /**
@@ -63,8 +64,10 @@ export class ReviewCommentService {
       return;
     }
 
-    rows[updateRowIndex] = CsvStructure.formatAsCsvLine(this.finalizeComment(comment));
-    setCsvFileLines(this.reviewFile, rows);
+    this.finalizeComment(comment, this.workspaceRoot).then((entry: CsvEntry) => {
+      rows[updateRowIndex] = CsvStructure.formatAsCsvLine(entry);
+      setCsvFileLines(this.reviewFile, rows);
+    });
   }
 
   async deleteComment(id: string, label: string) {
@@ -117,34 +120,32 @@ export class ReviewCommentService {
    * @param comment The comment to finalize
    * @return The finalized comment
    */
-  private finalizeComment(comment: CsvEntry): CsvEntry {
+  private async finalizeComment(comment: CsvEntry, workspaceRoot: string): Promise<CsvEntry> {
     const copy = { ...comment };
 
-    const gitDirectory = workspace.getConfiguration().get('code-review.gitDirectory') as string;
-    const gitRepositoryPath = path.resolve(this.workspaceRoot, gitDirectory);
-
     try {
-      copy.sha = gitCommitId({ cwd: gitRepositoryPath });
+      copy.revision = await revision(comment.filename, workspaceRoot);
     } catch (error) {
-      copy.sha = '';
-      console.log('Not in a git repository. Leaving SHA empty', error);
+      copy.revision = '';
+      window.showErrorMessage(`Repository not under version control as configured in the plugin's settings.
+      Leaving revision empty.\n\Details: ${error}`);
     }
 
-    const startAnker = startLineNumberFromStringDefinition(copy.lines);
-    const endAnker = endLineNumberFromStringDefinition(copy.lines);
-    copy.url = this.remoteUrl(copy.sha, copy.filename, startAnker, endAnker);
+    const startAnchor = startLineNumberFromStringDefinition(copy.lines);
+    const endAnchor = endLineNumberFromStringDefinition(copy.lines);
+    copy.url = this.remoteUrl(copy.revision, copy.filename, startAnchor, endAnchor);
 
     return copy;
   }
 
   /**
    * Build the remote URL
-   * @param sha a git SHA that's included in the URL
+   * @param revision a git revision that's included in the URL
    * @param filePath the relative file path
    * @param start the first line from the first selection
    * @param end the last line from the first selection
    */
-  private remoteUrl(sha: string, filePath: string, start?: number, end?: number) {
+  private remoteUrl(revision: string, filePath: string, start?: number, end?: number) {
     const customUrl = workspace.getConfiguration().get('code-review.customUrl') as string;
     const baseUrl = workspace.getConfiguration().get('code-review.baseUrl') as string;
 
@@ -154,15 +155,15 @@ export class ReviewCommentService {
       return '';
     } else if (customUrl) {
       return customUrl
-        .replace('{sha}', sha)
+        .replace('{revision}', revision)
         .replace('{file}', filePathWithoutLeadingAndTrailingSlash)
         .replace('{start}', start ? start.toString() : '0')
         .replace('{end}', end ? end.toString() : '0');
     } else {
       const baseUrlWithoutTrailingSlash = removeTrailingSlash(baseUrl);
-      const shaPart = sha ? `${sha}/` : '';
-      const ankerPart = start && end ? `#L${start}-L${end}` : '';
-      return `${baseUrlWithoutTrailingSlash}/${shaPart}${filePathWithoutLeadingAndTrailingSlash}${ankerPart}`;
+      const revPart = revision ? `${revision}/` : '';
+      const anchorPart = start && end ? `#L${start}-L${end}` : '';
+      return `${baseUrlWithoutTrailingSlash}/${revPart}${filePathWithoutLeadingAndTrailingSlash}${anchorPart}`;
     }
   }
 
